track living clients, avoid responses accumulating
When any Requester is connected, it will register itself
in an "active clients" set ("txscale.clients") and the server
will only push to a client response list if the client is registered
in the txscale.clients set.

also work around a race condition in crapc.py which would previously
cause the crapc client to try perform more protocol actions even
after the protocol was disconnected.